+++ vitrivr-engine-core/src/main/kotlin/org/vitrivr/engine/core/model/content/factory/InMemoryContentFactory.kt
@@ -1,10 +1,6 @@
 package org.vitrivr.engine.core.model.content.factory
 
-import org.vitrivr.engine.core.content.ContentFactory
-import org.vitrivr.engine.core.model.content.element.AudioContent
-import org.vitrivr.engine.core.model.content.element.ImageContent
-import org.vitrivr.engine.core.model.content.element.Model3DContent
-import org.vitrivr.engine.core.model.content.element.TextContent
+import org.vitrivr.engine.core.model.content.element.*
 import org.vitrivr.engine.core.model.content.impl.memory.InMemoryAudioContent
 import org.vitrivr.engine.core.model.content.impl.memory.InMemoryImageContent
 import org.vitrivr.engine.core.model.content.impl.memory.InMemoryMeshContent
@@ -13,6 +9,12 @@ import org.vitrivr.engine.core.model.mesh.Model3D
 import java.awt.image.BufferedImage
 import java.nio.ShortBuffer
 
+/**
+ * A [ContentFactory] that keeps [ContentElement]s in memory.
+ *
+ * @author Ralph Gasser
+ * @version 1.0.0
+ */
 class InMemoryContentFactory : ContentFactory {
 
     override fun newImageContent(bufferedImage: BufferedImage): ImageContent = InMemoryImageContent(bufferedImage)
