+++ vitrivr-engine-core/src/main/resources/META-INF/services/org.vitrivr.engine.core.model.content.factory.ContentFactory
